Better styling for tutorial selection page
It uses cards similar to Flight Launch and Flight Prime.

fight-common has been added as a dependency to gain access to its SASS
files. None of the components in flight-common can be readily used by
flight-tutorials as flight-common has a number of peer dependencies that
flight-tutorials does not use.  This is all solvable, but I'm further
inclined to find a way to replace flight-common with some other
component library.