Prompt to restart the session when it terminates
If the terminal session terminates, we now prompt the user for whether
they wish to restart the session.

This is very useful for running Flight Tutorials on clusters launched by
Flight Launch: the first time that a user logs on to the cluster they
will be prompted to change their password and then there session will be
terminated.  We now ask if the user wants to restart their session will
help step the user through what to do next.