flight-reactware is a peer dependency
Having multiple versions of `flight-reactware` installed causes issues
with `react-scroll` not scrolling to the correct location.  I'm not sure
if this is due to a clash in `react-scroll` versions, or simply due to
the possibility of multiple "instances" of `react-scroll` being included
in the bundle.

In either case, ensuring that `flight-tutorials-client` does not include
its own version of `flight-reactware` fixes the issue.