Ensure positive rows and columns
When the terminal is first rendered, it is possible for it to have a
negative height.  This is quickly changed by either CSS `height:
calc(...)` running or by `react-sizeme` running.

Prior to this commit, the negative height would result in a negative
number of rows being calculated.  This would cause the terminal to
break in a non recoverable way.

This commit addresses the issue by ensuring that the calculated rows
(and columns) are at least 1.